Silence warnings on linux hosts
svn path=/trunk/; revision=38582

+++ reactos/tools/rsym.c
@@ -489,7 +489,7 @@ ProcessRelocations(ULONG *ProcessedRelocsLength, void **ProcessedRelocs,
   *ProcessedRelocs = malloc(RelocSectionHeader->SizeOfRawData);
   if (NULL == *ProcessedRelocs)
     {
-      fprintf(stderr, "Failed to allocate %lu bytes for relocations\n", RelocSectionHeader->SizeOfRawData);
+      fprintf(stderr, "Failed to allocate %u bytes for relocations\n", (unsigned int)RelocSectionHeader->SizeOfRawData);
       return 1;
     }
   *ProcessedRelocsLength = 0;
@@ -565,7 +565,7 @@ CreateOutputFile(FILE *OutFile, void *InData,
   OutHeader = malloc(StartOfRawData);
   if (NULL == OutHeader)
     {
-      fprintf(stderr, "Failed to allocate %lu bytes for output file header\n", StartOfRawData);
+      fprintf(stderr, "Failed to allocate %u bytes for output file header\n", (unsigned int)StartOfRawData);
       return 1;
     }
   memset(OutHeader, '\0', StartOfRawData);
@@ -678,7 +678,7 @@ CreateOutputFile(FILE *OutFile, void *InData,
       PaddedRosSym = malloc(RosSymFileLength);
       if (NULL == PaddedRosSym)
         {
-          fprintf(stderr, "Failed to allocate %lu bytes for padded .rossym\n", RosSymFileLength);
+          fprintf(stderr, "Failed to allocate %u bytes for padded .rossym\n", (unsigned int)RosSymFileLength);
           return 1;
         }
       memcpy(PaddedRosSym, RosSymSection, RosSymLength);
